Made bets removable property into is_removable method taking the table as a parameter since for PassLine and DontPass it should be dependent on the table.
Got rid of place_odds from the bet. User should be getting the odds bet object and adding it to the player anyways.

Made a get_odds_bet abstract method for AllowsOdds to remove the odds_type parameter and simplify how that works allowing each AllowsOdds bet to get its own method, PassLine and DontPass can now be based on the table.

For PassLine and DontPass made it so that they are now dependent on the table to get their winning and losing numbers. This simplifies everything, but mostly testing as now you don't need to go through each pre-roll with the bet in order to get the correct "point" attribute. A PassLine bet on a table with a point of 8 should ALWAYS have winning number of 8. The old way allowed for this to not be the case.

Simplified players add_odds and made it work better. A player who is doing add_odds for PassLine shouldn't have to specify a number for the bet, it should just be based on the table. Now player can just do player.add_odds(10, [PassLine]) instead of specifying the number.